+++ vfs2provider-jdbctable/src/test/java/com/seeburger/vfs2/provider/jdbctable/test/StandaloneClient.java
@@ -12,18 +12,13 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.sql.SQLException;
-import java.text.Format;
-import java.text.MessageFormat;
-import java.util.Date;
 
 import javax.sql.DataSource;
 
 import org.apache.commons.vfs2.CacheStrategy;
-import org.apache.commons.vfs2.FileContent;
 import org.apache.commons.vfs2.FileObject;
 import org.apache.commons.vfs2.FileSystemException;
 import org.apache.commons.vfs2.FileSystemManager;
-import org.apache.commons.vfs2.FileType;
 import org.apache.commons.vfs2.Selectors;
 import org.apache.commons.vfs2.impl.DefaultFileSystemManager;
 import org.apache.derby.jdbc.EmbeddedDataSource;
@@ -86,11 +81,11 @@ public class StandaloneClient
             os.write(1); os.write(2); os.write(3); os.close();
             ax = manager.resolveFile(ax.toString());
             if (i % 200 == 0)
-                TreePrinter.printTree(ax ,String.format("(%3d) ", i), System.out);
+                TreePrinter.printTree(ax ,String.format("(%3d) ", Integer.valueOf(i)), System.out);
         }
 
         long middle = System.nanoTime();
-        System.out.printf("Write Time: %,.3f ms.%n", (middle - start) / 1000000.0);
+        System.out.printf("Write Time: %,.3f ms.%n", Double.valueOf((middle - start) / 1000000.0));
 
         System.out.println("Reading ("+count+") ...");
         for(int i=0; i<count; i++)
@@ -101,11 +96,11 @@ public class StandaloneClient
                 System.out.println("not 1 2 3");
             is.close();
             if (i % 200 == 0)
-                TreePrinter.printTree(ax ,String.format("(%3d) ", i), System.out);
+                TreePrinter.printTree(ax ,String.format("(%3d) ", Integer.valueOf(i)), System.out);
         }
 
         long end = System.nanoTime();
-        System.out.printf("Read Time: %,.3f ms. Have a good time.%n", (end - middle) / 1000000.0);
+        System.out.printf("Read Time: %,.3f ms. Have a good time.%n", Double.valueOf((end - middle) / 1000000.0));
 
         System.out.println("Destroying test files...");
         base.delete(Selectors.EXCLUDE_SELF);
